// src/app/lib/generated/models/UserInfoResponse.ts
export interface UserInfoResponse {
    /**
     * 
     * @type {string}
     * @memberof UserInfoResponse
     */
    userId: string;
    /**
     * 
     * @type {UserRole}
     * @memberof UserInfoResponse
     */
    role: UserRole;
    /**
     * 
     * @type {string}
     * @memberof UserInfoResponse
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UserInfoResponse
     */
    username: string;
    /**
     * 
     * @type {boolean}
     * @memberof UserInfoResponse
     */
    oauth: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UserInfoResponse
     */
    firstLogin: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof UserInfoResponse
     */
    feedbackGiven: boolean;
    /**
     * 
     * @type {Subscription}
     * @memberof UserInfoResponse
     */
    subscription: Subscription;
    /**
     * 
     * @type {Preferences}
     * @memberof UserInfoResponse
     */
    preferences: Preferences;
    /**
     * 
     * @type {boolean}
     * @memberof UserInfoResponse
     */
    hasUnseenImages: boolean;
}