// src/app/lib/generated/models/PostUserRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PostUserRequest
 */
export interface PostUserRequest {
    /**
     * The username of the user
     * @type {string}
     * @memberof PostUserRequest
     */
    username: string;
    /**
     * The email of the user
     * @type {string}
     * @memberof PostUserRequest
     */
    email: string;
    /**
     * The password of the user
     * @type {string}
     * @memberof PostUserRequest
     */
    password: string;
    /**
     * 
     * @type {boolean}
     * @memberof PostUserRequest
     */
    isDummy?: boolean | null;
}

/**
 * Check if a given object implements the PostUserRequest interface.
 */
export function instanceOfPostUserRequest(value: object): value is PostUserRequest {
    if (!('username' in value) || value['username'] === undefined) return false;
    if (!('email' in value) || value['email'] === undefined) return false;
    if (!('password' in value) || value['password'] === undefined) return false;
    return true;
}

export function PostUserRequestFromJSON(json: any): PostUserRequest {
    return PostUserRequestFromJSONTyped(json, false);
}

export function PostUserRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostUserRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'],
        'email': json['email'],
        'password': json['password'],
        'isDummy': json['is_dummy'] == null ? undefined : json['is_dummy'],
    };
}

export function PostUserRequestToJSON(json: any): PostUserRequest {
    return PostUserRequestToJSONTyped(json, false);
}

export function PostUserRequestToJSONTyped(value?: PostUserRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'email': value['email'],
        'password': value['password'],
        'is_dummy': value['isDummy'],
    };
}